Add necessary permissions
Internet for accessing Telegram servers.
Download for downloading shared files and caching stickers and profile pics.
Push for message notifications.
Recorder to record voice messages and to call (if you can compile libtgvoip on tizen).
Account Write is to add your Telegram account to the system Accounts section.
Location is for Location and Live Location features.